replaced sokol by SDL2
sokol_gfx is basically a totally new graphics API
to learn -- ain't got no time for that :) (And the
main selling point of sokol_gfx, independence of the
underlying API, is totally irrelevant here, as we're
tied to GLSL anyway).

So, let's bite the bullet and accept one DLL dependency
(SDL2) for Win32 builds for now. At least, we can make
the "out of box experience" somewhat nice by downloading
a local copy of the SDL2 development package during
the CMake run.