refactored FXAA shader
The original code was clearly written with very dumb,
non-optimizing compilers in mind: It's almost SSA form
(declaring temporary variables for almost all
intermediate computations), doesn't use vector operations
where appropriate, and the elephant in the room
is the 12 levels deeply nested if{} chain.

The if chain has been replaced by a simple loop,
and lots of intermediate computations have been
moved into the places where the results
are actually needed.

The end result is almost 70% shorter than
the previous version (which was already
a few thousands of lines shorter than
the original version).