make parameter comment parser more strict
If there was no comment directly after a parameter,
the parser would pick up the the next comment
as the parameter name, even if that's inside
some function.

This has been changed such that any non-symbolic
GLSL token after a uniform statement "cancels"
the current parameter, i.e. makes it impossible
to add a parameter comment to it.