using temporary VFS root when loading/saving pipelines from/to files
When saving a pipeline into a file, the file's parent directory
is temporarily added as a VFS root. This way, all paths to shaders
in (or below) that directory will be turned into relative paths.
The same is done when loading a pipeline file, so that these
relative paths can be resolved properly.

This is useful when preparing a pipeline that consists of
non-stock shaders -- it's then simply possible to share
a .gips file and a bunch of .glsl files, and importing that
again into GIPS should just work.